% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_based_functions_with_covariates.R
\name{AucPR_estimation}
\alias{AucPR_estimation}
\title{AucPR Estimation for Coefficient and Feature Selection}
\usage{
AucPR_estimation(
  df,
  X = NULL,
  y = "y",
  lambda,
  alpha = 1,
  c_to_use = "Youden",
  fold = NULL,
  regressors_betas = NULL,
  trace = 1,
  max.print = 10,
  ...
)
}
\arguments{
\item{df}{A data frame containing the complete dataset.}

\item{X}{A character vector of column names from `df` to be used as
regressor variables. Alternatively, a data frame containing only the
regressors. If not specified, all columns not in `y` are used.}

\item{y}{A character string specifying the column name of the target
variable in `df`. It must be a binomial variable with values 0 or 1.
Alternatively, a data frame containing only the target variable. Default
is "y".}

\item{lambda}{A single numeric value for the regularization parameter. A
larger value results in stronger penalization.}

\item{alpha}{The elastic-net mixing parameter. A value of 1 corresponds
to L1 penalization (Lasso), while a value between 0 and 1 corresponds
to Elastic-Net. Default is 1.}

\item{c_to_use}{character or numeric value specifying the classification
cut-point used to convert continuous predictions into binary outcomes. If set
to `"Youden"`, the function computes the optimal cut-point using Youden's Index.
If a numeric value is provided, it is used directly as the threshold.
Default is "Youden".}

\item{fold}{numeric. An optional fold number, used when the function is
called within a cross-validation loop. This is primarily for tracking
and reporting purposes. Default is `NULL`.}

\item{regressors_betas}{numeric vector. An optional vector containing the
"true" beta coefficients, if known. This is used to calculate additional
performance measures related to variable selection accuracy. Default is
`NULL`.}

\item{trace}{An integer controlling the amount of output. `0` for no
output, `1` for a summary of the results, and `2` for a detailed
breakdown of all steps. Default is 1.}

\item{max.print}{The number of elements to show when printing results.
Default is 10.}

\item{...}{Additional arguments passed to methods}
}
\value{
A list containing the optimal betas and a collection of performance
  metrics.
    \item{model_type}{A character string indicating the final model type used,
      either "AucPR_L1" or "AucPR_EN".}
    \item{betas_hat}{The estimated vector of optimal beta coefficients.}
    \item{X_model}{A character vector containing the names of all regressor
    variables (\code{X}) used in the model estimation. This vector defines
    the canonical order of the regressors used to align the coefficients
    in \code{betas_hat}.}
    \item{youden_index}{The Youden Index value at the optimal cut-point,
      which is a measure of the overall effectiveness of a diagnostic test.}
    \item{sensitivity}{The sensitivity (True Positive Rate) at the optimal
      cut-point.}
    \item{specificity}{The specificity (True Negative Rate) at the optimal
      cut-point.}
    \item{geometric_mean}{The geometric mean of sensitivity and
      specificity.}
    \item{fdr}{The False Discovery Rate at the optimal cut-point.}
    \item{mcc}{The Matthews Correlation Coefficient, a measure of the
      quality of binary classification.}
    \item{corrclass}{The overall correct classification rate.}
    \item{auc}{The Area Under the ROC Curve (AUC).}
    \item{lambda}{The penalization parameter used in the estimation.}
    \item{alpha}{The alpha parameter used in the estimation.}
    \item{c_hat}{The optimal cut-point for the `z_hat` score.}
    \item{z_hat}{A data frame with the estimated linear combination of
      regressors for each observation.}
    \item{y_hat}{A data frame with the predicted binary outcomes (0 or 1)
      for each observation.}
    \item{n_betas}{The number of non-zero beta coefficients estimated.}
    \item{n_total_var}{The total number of regressors considered.}
    \item{n_predicted_zeros}{The number of estimated betas that are zero.}
    \item{n_predicted_non_zeros}{The number of estimated betas that are
      non-zero.}
    \item{n_caught_betas}{The number of true non-zero betas correctly
      identified as non-zero (if `regressors_betas` is provided).}
    \item{n_non_caught_betas}{The number of true non-zero betas incorrectly
      identified as zero.}
    \item{n_caught_zero}{The number of true zero betas correctly identified
      as zero.}
    \item{n_zero_not_caught}{The number of true zero betas incorrectly
      identified as non-zero.}
    \item{estimation_time}{The time taken for the estimation, in minutes.}
}
\description{
function to estimate the optimal value of betas using the
AucPR method.

This function estimates the optimal regression coefficients
(betas) for a binary classification problem by maximizing the penalized
Area Under the Precision-Recall Curve (AUC-PR). The penalization is either
L1 (Lasso) or Elastic-Net, which helps in feature selection and regularization.
The function returns the estimated coefficients along with various
performance metrics.
}
\examples{
library(pye)
sim_data <- create_sample_with_covariates(
    rows_train = 100, cols = 40, cols_cov = 12, max_rho = 0.3, seed = 1)
df <- sim_data$train_df_scaled
X <- sim_data$X
y <- sim_data$y
regressors_betas <- sim_data$nregressors
alpha <- 1
lambda <- 0.1

AucPR_result <- AucPR_estimation(df = df, X = X, y = y, lambda = lambda, alpha = alpha,
        regressors_betas = regressors_betas, trace = 1)

print(AucPR_result)

}
