% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_lambda_for_penalized_models.R
\name{calibrate_lambda_min}
\alias{calibrate_lambda_min}
\title{Calibrate Minimum Value for the Penalty Parameter}
\usage{
calibrate_lambda_min(
  function_to_run,
  var_to_check = "betas_hat",
  lambda_start = 1e-04,
  lambda_inf = 1e-20,
  max_var = 1000,
  factor = 2
)
}
\arguments{
\item{function_to_run}{Function that needs to be cross-validated. It needs to
have a have just one parameter to change (theoretically called lambda),
changing which the number of selected variables might change.}

\item{var_to_check}{Name of the variable (a vector) to check if, reducing
lambda, the number of its elements increases.}

\item{lambda_start}{Starting value of lambda. If left too small it might take
longer (default is 0.0001)}

\item{lambda_inf}{floor value for lambda (default is 1e-20)}

\item{max_var}{Maximum number of variable to accept the result of the
function (default is 100.000)}

\item{factor}{Number to multiply or divide the considered lambda in the
step. The higher, the slower to find the optimal value. Default is 2.}
}
\value{
The first encountered value of the parameter (lambda inside the
function) at which the number of elements of var_to_check is lower then
max_var.
}
\description{
Function to calibrate the last used value of lambda, to make
cross-validation more effective. In particular, this function aim to solve
the problem that every model has its own "best" range of lambdas at which
the parameters are different from 0 and performs better. The function stops
where the number of elements of var_to_check is greater then 0.
}
\examples{
library(pye)
simMicroarrayData_cov02_dim50_covariates <- create_sample_with_covariates(
    rows_train = 100, cols = 40, cols_cov = 12, max_rho = 0.3, seed = 1)
df <- simMicroarrayData_cov02_dim50_covariates$train_df_scaled
#create the wrapper
penalty <- "L1"
wrapper <- function(lambda) {
   return(pye_KS_estimation(df = df, penalty = penalty, trace = 1, lambda = lambda,
   c_zero_fixed = FALSE, max_iter = 5))
}
#find the lambda min
var_to_check <- paste0("betas_hat_", penalty)
#accept maximum 100 betas
max_var <- 20
lambda_min <- calibrate_lambda_min (function_to_run = wrapper,
  var_to_check = var_to_check, lambda_start = 0.2, max_var = max_var)
print(lambda_min)

}
