% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study_other_models_with_covariates.R
\name{model_simulation_study_real_data}
\alias{model_simulation_study_real_data}
\title{model_simulation_study_real_data}
\usage{
model_simulation_study_real_data(
  n = 1000,
  df,
  X = names(df[, !(names(df) \%in\% c(y, C))]),
  y = "y",
  C = NULL,
  model_estimation_function,
  model_prediction_function,
  model_type,
  lambda,
  tau = 0,
  trace = 2,
  alpha = 0.5,
  regressors_betas = NULL,
  used_cores = 1,
  scaling = FALSE,
  c_function_of_covariates = FALSE,
  alpha_g = 0.5,
  penalty_g = "L1",
  kernel_g = "gaussian",
  a1_g = 3.7,
  a2_g = 3,
  trend_g = "monotone",
  gamma_start_input = NULL,
  gamma_start_default = "zeros",
  regressors_gammas = NULL,
  max_iter_g = 10000,
  delta_g = 1e-05,
  max_alpha_g = 10000,
  stepsizeShrink_g = 0.8,
  min_alpha_g = 1e-12,
  convergence_error_g = 1e-07,
  run_aauc = FALSE
)
}
\arguments{
\item{n}{number of esperiments. Default is 1000}

\item{df}{the input dataset}

\item{X}{regressors to consider in the estimation. It can be of type
dataframe, containing also the same name of the regressors included in df,
of just a vector of character. Default is all not present in y and C}

\item{y}{the target variable. It can be only binomial 0,1. It can be of type
dataframe, containing also the same name of the same target variable included
in df, or just a character. Default is "y".}

\item{C}{covariate variables. It can be of type dataframe, containing the
same covariates included in df, or just a vector of character. Default is NULL}

\item{model_estimation_function}{function to use for the estimation. The user
can choose among "glmnet_estimation", "all_svm_estimation", "AucPR_estimation"}

\item{model_prediction_function}{function to use for the prediction. The user
can choose among "glmnet_predict", "all_svm_predict", "AucPR_predict"}

\item{model_type}{model to use in the estimation. The user can choose
among "logLasso", "logElasticNet", "logSCAD", "logMCP", "SCADSVM",
"ElasticSCADSVM", "l1SVM", "enSVM", "AucPR_L1", "AucPR_EN"}

\item{lambda}{the penalization parameter of the regressors X}

\item{tau}{the penalization parameter of the covariates C, in covYI. Default 0,
i.e. no penalization term}

\item{trace}{2:visualize all the steps, 1:visualize just the result,
0:visualize nothing. Default is 1}

\item{alpha}{parameter for the Elastic-Net penalization term. Default is 0.5}

\item{regressors_betas}{a vector containing the real betas (if known). Default
is NULL, i.e. we do not know the real regressors}

\item{used_cores}{number of core used for the parallelization of the
process. if equal to 1, then no parallelization is adopted. Default is 1.}

\item{scaling}{if TRUE, the dataset is scaled. FALSE otherwise, Default is
FALSE.}

\item{c_function_of_covariates}{if TRUE, covYI is used to estimate the
cut-off point as function of the convariate information. If FALSE, the
covariate information is ignored. Default is FALSE}

\item{alpha_g}{parameter for the Elastic-Net penalization term in covYI.
Default is 0.5}

\item{penalty_g}{the considered penalty of covYI. To be chosen among L12, L1,
EN, SCAD and MCP. Default if "L1"}

\item{kernel_g}{the kernel type to use for the estimation of the density
function (tested only for "gaussian") in covYI.  Default is "gaussian"}

\item{a1_g}{parameter for the SCAD and MCP penalization term in covYI. Default is 3.7}

\item{a2_g}{parameter for the MCP penalization term in covYI. Default is 3.0}

\item{trend_g}{for covYI. If "monotone", mmAPG is used, if "nonmonotone",
mnmAPG is used. Default is "monotone"}

\item{gamma_start_input}{vector of a specific starting point for gammas.
Default is NULL, i.e. no input vector}

\item{gamma_start_default}{set the default starting point of gamma.
If "zeros", it starts with all zero values, if "corr" it starts with the
value of the correlation of every regressor with the target variable. Default
is "zeros"}

\item{regressors_gammas}{a vector containing the real gammas (if known).
Default is NULL}

\item{max_iter_g}{maximum number of iterations in the algorithms mmAPG and
mnmAPG in covYI. Default is 500}

\item{delta_g}{parameter for the convergence condition of the optimization
algorithm of covYI. Default is 1e-5}

\item{max_alpha_g}{maximum value of the step-parameter alpha in covYI.
Default is 100}

\item{stepsizeShrink_g}{parameter to adjust the step-size in the backtracking
line-search, in the optimization of covYI. Taking values between 0 and 1,
the closer to 1, the more accurate the estimation will be, the longer it
will take and viceversa. Default is 0.8}

\item{min_alpha_g}{minimum value of the step-parameter alpha in covYI.
Default is 1e-12}

\item{convergence_error_g}{in covYI, it is error to accept for considering the algorithm
converged. Default is 1e-5}

\item{run_aauc}{if FALSE the aAUC and aYI are not computed, to save stimation
time if not requested. Default is FALSE}
}
\value{
a list containing the classification measure related to every simulation.
}
\description{
function to perform the simulation on n different
experiments, splitting the dataset randomly in train and test,
estimating the considered competitor method of pye (penalized
logistic regressions, penalized SVM and penalized AUC-based
method), in the train and evaluating the classification
measures on the test set.
}
\examples{
library(pye)
cols <- 2000
cols_cov <- 20
seed=1
simMicroarrayData_cov02_dim50_covariates <- create_sample_with_covariates(
		rows_train=50, cols=cols, cols_cov=cols_cov, covar=0.2, seed=seed)
df <- simMicroarrayData_cov02_dim50_covariates$train_df_scaled
X <- simMicroarrayData_cov02_dim50_covariates$X
y <- simMicroarrayData_cov02_dim50_covariates$y
C <- simMicroarrayData_cov02_dim50_covariates$C
regressors_betas<-simMicroarrayData_cov02_dim50_covariates$nregressors
regressors_gammas<-simMicroarrayData_cov02_dim50_covariates$ncovariates
alpha <- 0.5
used_cores <- 1
n <- 5 #number of simulations
c_function_of_covariates <- TRUE

used_penalty_covYI <- c("L12", "L1") #c("L12", "L1", "EN", "SCAD", "MCP")
for (p in used_penalty_covYI){

  #--------------------------------------------------------------------------
  #                                  START                                  |
  #--------------------------------------------------------------------------
  other_models_vector = c("logLasso", "SCADSVM", "AucPR_L1") # c("logLasso",
						# "logElasticNet", "logSCAD", "logMCP",
                        # "SCADSVM", "ElasticSCADSVM", "l1SVM", "enSVM",
                        # "AucPR_L1", "AucPR_EN")

  functions <- c("glmnet", "all_svm", "AucPR")

  measures <- c("ccr") #c("auc", "aauc", "aYI", "ccr", "yi", "gm")
  for (m in other_models_vector){
    if (m \%in\% c("logLasso", "logElasticNet", "logSCAD", "logMCP")){f <- functions[1]}
    if (m \%in\% c("SVM_linear", "SVM_polynomial", "SVM_radial", "SVM_sigmoid", "SCADSVM",
					"ElasticSCADSVM", "l1SVM", "enSVM")){f <- functions[2]}
    if (m \%in\% c("AucPR_L1", "AucPR_EN")){f <- functions[3]}
    for (mm in 1:length(measures)){
      cat("---> Starting with measure", measures[mm], ", for the model", m, "<---\n")
      name_est <- get(paste0(f ,"_estimation"))
      name_pred <- get(paste0(f ,"_predict"))
      name <- paste0(m, "_covYI_", p, "_sim_study_" , measures[mm])
      lambda_to_use <- 0.1
      if (c_function_of_covariates==TRUE){
        tau_to_use <- 0.1
      } else {
        tau_to_use <- 0
      }
      assign(name, model_simulation_study_real_data(n=n, df=df, X=X, y=y, C=C,
         model_estimation_function=name_est,
         model_prediction_function=name_pred,
         model_type=m,
         trace=1,
         lambda=lambda_to_use,
         tau=tau_to_use,
         alpha=alpha,
         alpha_g=0.5,
         penalty_g=p,
         used_cores=used_cores,
         c_function_of_covariates=c_function_of_covariates,
			max_iter_g=10 #because it is an example. Please use more iterations
					   ))

      cat("Next measure!")
    }
    cat("Next model!")
  }
  cat("Next penalty!")
}
cat("Computation finished! Start saving data.")
print(name)

}
