% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_dataset_creation.R
\name{PBC_Mayo_Clinic_data_for_longpye}
\alias{PBC_Mayo_Clinic_data_for_longpye}
\title{PBC_data_load}
\usage{
PBC_Mayo_Clinic_data_for_longpye(T = 4)
}
\arguments{
\item{T}{latest time to consider, i.e. from 1 to T. 4 is the default because
we already excluded the subjects with less than 4 visits.}
}
\value{
the Mayo Clinic Primary Biliary Cholangitis Data (PBC) ready to be
used in longpye
}
\description{
function to load the Mayo Clinic Primary Biliary Cholangitis
Data (PBC) ready to be used in longpye.
}
\examples{
library(pye)
df <- PBC_Mayo_Clinic_data_for_longpye()

}
