% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_proximal_operators.R
\name{proximal_operator_EN}
\alias{proximal_operator_EN}
\title{Proximal Operator for the Elastic-Net Penalty}
\usage{
proximal_operator_EN(betas, lambda, alpha, ...)
}
\arguments{
\item{betas}{\code{numeric} vector to which the proximal operator should
be applied.}

\item{lambda}{\code{numeric}. The overall non-negative penalization parameter
\eqn{\lambda}.}

\item{alpha}{\code{numeric}. The mixing parameter \eqn{\alpha \in [0, 1]}.
\eqn{\alpha = 1} is pure \eqn{\textrm{L}_{1}} (Lasso); 
\eqn{\alpha = 0} is pure \eqn{\textrm{L}_2} (Ridge).}

\item{...}{Additional arguments passed to methods}
}
\value{
A \code{numeric} vector of the same length as \code{betas},
  containing the result of applying the Elastic-Net proximal operator.
}
\description{
Computes the proximal operator for the Elastic-Net penalty,
  which is a convex combination of \eqn{\textrm{L}_{1}} (Lasso) and
  \eqn{\textrm{L}_2} (Ridge). This operator serves as the backward step in
  Proximal Gradient optimization.
}
\details{
The Elastic-Net penalty is a convex combination of the \eqn{\textrm{L}_{1}}
and \eqn{\textrm{L}_2} penalties. The combined penalty \eqn{P(\beta)} is:
\deqn{P(\beta) = \lambda \left( \alpha |\beta| + (1-\alpha) \frac{1}{2} \beta^2 \right)}
The proximal operator for the Elastic-Net penalty has a closed-form solution:
\deqn{\text{prox}_{P}(\beta) = \frac{1}{1 + \lambda (1-\alpha)} \cdot \text{sign}(\beta) \cdot \max(0, |\beta| - \lambda \alpha)}
This function applies this operator element-wise to the input \code{betas} vector.
}
\examples{
library(pye)
betas <- seq(-2, 2, by = 0.2)
lambda <- 0.5
alpha <- 0.5
proximal_operator_EN(betas = betas, lambda = lambda, alpha = alpha)

# Equivalent to Lasso
prox_lasso <- proximal_operator_EN(betas = betas, lambda = 0.5, alpha = 1)

# Equivalent to a scaled Ridge
prox_ridge <- proximal_operator_EN(betas = betas, lambda = 0.5, alpha = 0)

}
