% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_proximal_operators.R
\name{proximal_operator_L1}
\alias{proximal_operator_L1}
\title{Proximal Operator for the Convex \eqn{\textrm{L}_{1}} (Lasso) Penalty}
\usage{
proximal_operator_L1(betas, lambda, ...)
}
\arguments{
\item{betas}{\code{numeric} vector to which the proximal operator should be applied.}

\item{lambda}{\code{numeric}. The non-negative penalization parameter \eqn{\lambda}.}

\item{...}{Additional arguments passed to methods}
}
\value{
A \code{numeric} vector of the same length as \code{betas},
  containing the result of applying the soft-thresholding operator.
}
\description{
Computes the proximal operator for the convex \eqn{\textrm{L}_{1}} (Lasso)
  penalty. This operator is the soft-thresholding operator and serves
  as the backward step in Proximal Gradient optimization.
}
\details{
The \eqn{\textrm{L}_{1}} penalty, also known as the Lasso penalty, is a convex
penalty commonly used to enforce sparsity (set coefficients to zero).
Its proximal operator has a closed-form solution known as the
soft-thresholding operator:

\deqn{\text{prox}_{\lambda \textrm{L}_1}(\beta) = \text{sign}(\beta) \cdot \max(0, |\beta| - \lambda)}

This function applies this operator element-wise to the input \code{betas} vector.
}
\examples{
# Example usage:
betas <- seq(-2, 2, by = 0.2)
lambda <- 0.5
proximal_operator_L1(betas = betas, lambda = lambda)

# Another example with different lambda
lambda2 <- 0.1
proximal_operator_L1(betas = betas, lambda = lambda2)

}
