% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_proximal_operators.R
\name{proximal_operator_MCP}
\alias{proximal_operator_MCP}
\title{Proximal Operator for the Non-Convex MCP Penalty}
\usage{
proximal_operator_MCP(betas, lambda, a, ...)
}
\arguments{
\item{betas}{\code{numeric} vector to which the proximal operator should be applied.}

\item{lambda}{\code{numeric}. The non-negative penalization parameter \eqn{\lambda}.}

\item{a}{\code{numeric}. The hyperparameter \eqn{a > 1} for the MCP penalty.
It controls the threshold after which the penalty is zero. $a=3.0$ is a common choice.}

\item{...}{Additional arguments passed to methods}
}
\value{
A \code{numeric} vector of the same length as \code{betas},
  containing the result of applying the MCP proximal operator.
}
\description{
Computes the proximal operator for the non-convex Minimax
  Concave Penalty (MCP). This operator serves as the backward step
  in Proximal Gradient optimization.
}
\details{
The MCP penalty is a non-convex penalty that smoothly transitions from the
\eqn{\textrm{L}_{1}} penalty (for small coefficients) to zero (for large coefficients),
thereby reducing estimation bias. Its proximal operator is defined piecewise:

\enumerate{
\item If \eqn{|\beta| \le \lambda}:
  \deqn{\text{prox}_{\lambda P}(\beta) = 0}
\item If \eqn{\lambda < |\beta| \le a\lambda}:
  \deqn{\text{prox}_{\lambda P}(\beta) = \frac{\beta - \text{sign}(\beta)\lambda}{1 - 1/a}}
\item If \eqn{|\beta| > a\lambda}:
  \deqn{\text{prox}_{\lambda P}(\beta) = \beta}
}

This function applies this operator element-wise to the input \code{betas} vector.
}
\examples{
# Example usage:
betas <- seq(-2, 2, by = 0.2)
lambda <- 0.5
a <- 3.0 # Common choice for 'a'
prox_betas <- proximal_operator_MCP(betas = betas, lambda = lambda, a = a)
print(prox_betas)

# Example with a different 'a' value
prox_betas2 <- proximal_operator_MCP(betas = betas, lambda = 0.3, a = 2)
print(prox_betas2)

}
