% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_proximal_operators.R
\name{proximal_operator_SCAD}
\alias{proximal_operator_SCAD}
\title{Proximal Operator for the Non-Convex SCAD Penalty}
\usage{
proximal_operator_SCAD(betas, lambda, a, ...)
}
\arguments{
\item{betas}{\code{numeric} vector to which the proximal operator should be applied.}

\item{lambda}{\code{numeric}. The non-negative penalization parameter \eqn{\lambda}.}

\item{a}{\code{numeric}. The hyperparameter \eqn{a > 2} for the SCAD penalty.
It controls the concavity of the penalty function. \eqn{a=3.7} is a common choice.}

\item{...}{Additional arguments passed to methods}
}
\value{
A \code{numeric} vector of the same length as \code{betas},
  containing the result of applying the SCAD proximal operator.
}
\description{
Computes the proximal operator for the non-convex Smoothly
  Clipped Absolute Deviation (SCAD) penalty. This operator serves as the
  backward step in Proximal Gradient optimization.
}
\details{
The SCAD penalty is a non-convex penalty designed for sparse and unbiased
estimation. Its proximal operator is defined piecewise based on \eqn{\lambda} and $a$:

\enumerate{
\item If \eqn{|\beta| \le 2\lambda}:
  \deqn{\text{prox}_{\lambda P}(\beta) = \text{sign}(\beta) \cdot \max(0, |\beta| - \lambda)}
\item If \eqn{2\lambda < |\beta| \le a\lambda}:
  \deqn{\text{prox}_{\lambda P}(\beta) = \frac{(a - 1)\beta - \text{sign}(\beta)a\lambda}{a-2}}
\item If \eqn{|\beta| > a\lambda}:
  \deqn{\text{prox}_{\lambda P}(\beta) = \beta}
}

This function applies this operator element-wise to the input \code{betas} vector.
}
\examples{
# Example usage:
betas <- seq(-2, 2, by = 0.2)
lambda <- 0.5
a <- 3.7 # Common choice for 'a'
prox_betas <- proximal_operator_SCAD(betas = betas, lambda = lambda, a = a)
print(prox_betas)

# Example with a different 'a' value
prox_betas2 <- proximal_operator_SCAD(betas = betas, lambda = 0.3, a = 3)
print(prox_betas2)

}
