% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalizedSVM_with_covariates.R
\name{psvm_compute_cv}
\alias{psvm_compute_cv}
\title{Cross-Validation for Optimal Penalized svm Regularization Parameter
Selection}
\usage{
psvm_compute_cv(
  n_folds,
  df,
  X = NULL,
  y = "y",
  C = NULL,
  lambda,
  tau = 0,
  w_g = 0.5,
  c_to_use = NULL,
  regressors_betas = NULL,
  trace = 1,
  alpha = 0.5,
  model_type = c("SCADSVM", "ElasticSCADSVM", "l1SVM", "enSVM"),
  seed = 1,
  used_cores = 1,
  scaling = FALSE,
  c_function_of_covariates = FALSE,
  simultaneous = FALSE,
  measure_to_select_lambda = "ccr",
  alpha_g = 0.5,
  penalty_g = "L1",
  kernel_g = "gaussian",
  a1_g = 3.7,
  a2_g = 3,
  trend_g = "monotone",
  gamma_start_input = NULL,
  gamma_start_default = "zeros",
  regressors_gammas = NULL,
  max_iter_g = 10000,
  delta_g = 1e-05,
  max_alpha_g = 10000,
  stepsizeShrink_g = 0.8,
  min_alpha_g = 1e-12,
  convergence_error_g = 1e-07,
  run_aauc = FALSE,
  log_file = "log_SVM_models.txt"
)
}
\arguments{
\item{n_folds}{Integer. The number of folds to use for cross-validation
(must be >= 2).}

\item{df}{Data frame. The input dataset containing the target variable,
regressors, and covariates.}

\item{X}{Character vector. Names of the regressor variables (features).
Defaults to all columns in "df" that are not "y" or "C". Can also be a
data frame containing only the regressors.}

\item{y}{Character. Name of the binary target variable (outcome), coded as
0 and 1. Defaults to "y". Can also be a data frame containing only the
target variable.}

\item{C}{Character vector. Names of the covariate variables for
covariate-adjusted cut-point estimation. Defaults to "NULL" (no
covariates). Can also be a data frame containing only the covariates.}

\item{lambda}{Numeric vector. The regularization parameter(s) for the
regressors "X".}

\item{tau}{Numeric vector. The regularization parameter(s) for the
covariates "C" in "covYI". If "c_function_of_covariates = TRUE", this
parameter cannot be "NULL" or contain only zeros. Default is 0 (no
penalization).}

\item{w_g}{A `numeric` value between 0 and 1 specifying the weight for the 
Weighted Youden Index in covYI. Sensitivity is weighted by `w_g` and specificity by 
`1 - w_g`. Default is 0.5 (which corresponds to the standard Youden Index).}

\item{c_to_use}{Character or numeric. Specifies the classification
cut-point used to convert continuous predictions into binary outcomes. If
set to "Youden", the function computes the optimal cut-point using
Youden's Index. If a numeric value is provided, it is used directly as 
the threshold. If "NULL", a default is assigned based on 
"model_type": "SCADSVM" uses "0", "ElasticSCADSVM", "l1SVM" or "enSVM" 
use "Youden".}

\item{regressors_betas}{numeric vector. An optional vector containing the
"true" beta coefficients, if known. This is used to calculate additional
performance measures related to variable selection accuracy. Default is
"NULL".}

\item{trace}{Integer. Level of verbosity: 0 = silent, 1 = brief summary, 2
= verbose. Default is 1.}

\item{alpha}{numeric in (0, 1]. Elastic-net mixing parameter for sparseSVM
(alpha = 1 corresponds to L1). Default is 0.5.}

\item{model_type}{Character. The SVM model to use. Options: "SCADSVM",
"ElasticSCADSVM", "l1SVM", "enSVM".}

\item{seed}{Integer. Random seed for reproducibility. Default is 1.}

\item{used_cores}{Integer. Number of cores to use for parallel processing.
If 1, no parallelization is used. Default is 1.}

\item{scaling}{Logical. If "TRUE", the dataset is scaled. Default is
"FALSE".}

\item{c_function_of_covariates}{Logical. If "TRUE", "covYI" is used to
estimate the cut-off point as a function of the covariates. Default is
"FALSE".}

\item{simultaneous}{Logical. If "TRUE" (and
"c_function_of_covariates = TRUE"), gammas are estimated simultaneously
with betas. If "FALSE", gammas are estimated as a second step. Default
is "FALSE".}

\item{measure_to_select_lambda}{Character. The performance measure used to
select the best lambda when "simultaneous = FALSE". Options: "auc",
"aauc", "aYI", "ccr", "yi", "gm", "fdr", "mcc", "sen",
"spc". Default is "ccr" (correct classification rate).}

\item{alpha_g}{Numeric. Elastic-Net mixing parameter for "covYI". Default
is 0.5.}

\item{penalty_g}{Character. The penalty of "covYI". Options: "L12",
"L1", "EN", "SCAD", "MCP". Default is "L1".}

\item{kernel_g}{Character. Kernel type for density estimation in "covYI".
Default is "gaussian".}

\item{a1_g}{Numeric. Parameter for the SCAD and MCP penalties in "covYI".
Default is 3.7.}

\item{a2_g}{Numeric. Parameter for the MCP penalty in "covYI". Default is
3.0.}

\item{trend_g}{Character. For "covYI", if "monotone", mmAPG is used; if
"nonmonotone", mnmAPG is used. Default is "monotone".}

\item{gamma_start_input}{Numeric vector. A specific starting point for
gammas. Default is "NULL".}

\item{gamma_start_default}{Character. Sets the default starting point of
gamma. If "zeros", it starts with all zero values; if "corr", it
starts with the value of the correlation of every regressor with the
target variable.}

\item{regressors_gammas}{Numeric vector. Optional vector containing the true
gammas (if known).}

\item{max_iter_g}{Integer. Maximum number of iterations in the algorithms
mmAPG and mnmAPG in "covYI". Default is 10000.}

\item{delta_g}{Numeric. Parameter for the convergence condition of the
optimization algorithm of "covYI". Default is 1e-5.}

\item{max_alpha_g}{Numeric. Maximum value of the step-parameter alpha in
"covYI". Default is 100.}

\item{stepsizeShrink_g}{Numeric. Parameter to adjust the step-size in the
backtracking line-search, in the optimization of "covYI". Taking values
between 0 and 1, the closer to 1, the more accurate the estimation will
be, the longer it will take and vice versa. Default is 0.8.}

\item{min_alpha_g}{Numeric. Minimum value of the step-parameter alpha in
"covYI". Default is 1e-12.}

\item{convergence_error_g}{Numeric. Error to accept for considering the
algorithm converged in "covYI". Default is 1e-7.}

\item{run_aauc}{Logical. If "FALSE", the aAUC and aYI are not computed, to
save estimation time if not requested. Default is "FALSE".}

\item{log_file}{Character. Path to a file for logging output from parallel
workers. If "NULL", output goes to the console.
Defaults to "log_SVM_models.txt".}
}
\value{
A list containing the optimal values of lambda (and possibly tau) to
  estimate betas (and possibly gammas) and the value of the main accuracy
  measure for all the folds. The list includes:
  \item{model_type}{The SVM model type used.}
  \item{c_to_use}{The classification cut-point used.}
  \item{penalty_g}{The penalty type used for covYI.}
  \item{cv_time}{The time taken for cross-validation.}
  \item{auc_first_step, aauc_first_step, aYI_first_step, 
    youden_index_first_step, sensitivity_first_step,
    specificity_first_step, geometric_mean_first_step, fdr_first_step, 
    mcc_first_step, corrclass_first_step}{A series of nested lists,
     each containing performance matrices for the respective measure
     of the first-step estimated method.}
  \item{auc, aauc, aYI, youden_index, sensitivity,
    specificity, geometric_mean, fdr, mcc, corrclass}{A series of
    nested lists, each containing performance matrices for the 
    respective measure of the final estimated method.}
  \item{lambda_hat_yi, lambda_hat_auc, lambda_hat_aauc, lambda_hat_aYI,
    lambda_hat_ccr, lambda_hat_sen, lambda_hat_spc, lambda_hat_gm}{The
    optimal lambda values selected based on different performance
    measures.}
  \item{tau_hat_yi, tau_hat_auc, tau_hat_aauc, tau_hat_aYI, tau_hat_ccr,
    tau_hat_sen, tau_hat_spc, tau_hat_gm}{The optimal tau values
    selected based on different performance measures (if
    `c_function_of_covariates = TRUE`).}
  \item{c_function_of_covariates}{Indicates whether the cut-off point was
    estimated as a function of covariates.}
  \item{simultaneous}{Indicates whether betas and gammas were estimated
    simultaneously.}
  \item{measure_to_select_lambda}{The measure used to select lambda.}
  \item{lambda_star}{The optimal lambda value when using a sequential
    search.}
  \item{n_betas}{The number of non-zero beta coefficients.}
  \item{n_betas_star}{The number of non-zero beta coefficients when using
    lambda_star.}
  \item{n_gammas}{The number of non-zero gamma coefficients.}
  \item{betas_star}{The estimated beta coefficients when using
    lambda_star.}
  \item{betas}{The estimated beta coefficients.}
  \item{gammas}{The estimated gamma coefficients.}
}
\description{
Performs k-fold cross-validation for penalized SVM models,
  optionally incorporating covariate-adjusted cut-point optimization via
  the covYI method. This function estimates the optimal values of the
  regularization parameters lambda (for feature selection) and tau (for
  covariate adjustment of the classification threshold).
}
\examples{
library(pye)
# Generate a sample dataset
sim_data <- create_sample_with_covariates(
  rows_train = 100, cols = 40, cols_cov = 12, max_rho = 0.3, seed = 1)
df <- sim_data$train_df_scaled
X <- sim_data$X
y <- sim_data$y
C <- sim_data$C
regressors_betas <- sim_data$nregressors
regressors_gammas <- sim_data$ncovariates

# Define parameters
lambda <- c(0.1, 0.05)
c_function_of_covariates <- FALSE
model_type <- "SCADSVM"

# Run cross-validation
result <- psvm_compute_cv(n_folds = 2,
                          df = df,
                          X = X,
                          y = y,
                          C = C,
                          lambda = lambda,
                          regressors_betas = regressors_betas,
                          regressors_gammas = regressors_gammas,
                          c_function_of_covariates = c_function_of_covariates,
                          model_type = model_type
)

# Print the results
print(result$model_type)
print(result$lambda_hat_ccr)

}
