% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_df_for_pye.R
\name{scaling_df_for_pye}
\alias{scaling_df_for_pye}
\title{Center and Scale Continuous Regressors for pye Modeling}
\usage{
scaling_df_for_pye(df, X, y)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} containing all variables.}

\item{X}{A \code{character} vector of column names considered as potential
regressors. Only numeric variables from this list will be scaled.}

\item{y}{A \code{character} string: the name of the target variable.
This variable will not be scaled.}
}
\value{
A \code{list} containing two elements:
  \item{df_scaled}{A \code{data.frame} with the selected numeric
    regressors scaled (centered to mean 0 and scaled to SD 1).}
  \item{scaling_params}{A \code{list} containing:
    \itemize{
      \item \code{original_mu}: Means of the variables that were scaled.
      \item \code{original_stdev}: Standard deviations of the variables.
    }
    These parameters enable consistent scaling of new data.}
}
\description{
Scales and centers specified numeric variables in a data
  frame (mean 0, SD 1). It automatically excludes the target variable
  (\code{y}) and variables identified as binary (dummy), ensuring only
  continuous regressors are standardized. This prepares data for models
  where standardization is necessary, such as the pye optimization.
}
\details{
The function first identifies potential regressors from the \code{X} argument.
It then excludes variables that are the target (\code{y}), are not
numeric, or are identified as binary (having exactly two unique values).
Only the remaining continuous numeric variables are scaled.
The scaling parameters are returned to allow for the consistent
transformation of future datasets (e.g., test or new data).
}
\examples{
library(pye)

# Simulate the dataframe
sim_data <- create_sample_with_covariates(
  rows_train = 100, rows_test = 50, cols = 40, cols_cov = 12, max_rho = 0.3, seed = 1)
df <- sim_data$df
X_cols_name <- sim_data$X
y_col_name <- sim_data$y
C_cols_name <- sim_data$C

# Now call the scaling function
scaled_output <- scaling_df_for_pye(
  df = df,
  X = c(X_cols_name, C_cols_name),
  y = y_col_name
)

print(head(scaled_output$df_scaled))
print(scaled_output$scaling_params$original_mu)
print(scaled_output$scaling_params$original_stdev)

}
