% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{create_lambda}
\alias{create_lambda}
\title{create_lambda}
\usage{
create_lambda(n = 100, lmax = 10, lmin = 1)
}
\arguments{
\item{n}{Number of lambdas}

\item{lmax}{Maximum lambda value}

\item{lmin}{Minimum lambda value}
}
\value{
A a vector containing n chosen values of lambdas
}
\description{
Function to create a sequence of n lambdas, given a lambda-max and a lambda-min
}
\examples{
lambdas <- create_lambda(n=100, lmax=10, lmin=1)
}
