% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_dataset_creation.R
\name{create_sample}
\alias{create_sample}
\title{create_sample}
\usage{
create_sample(
  rows_train = 50,
  cols = 2000,
  cov = 0.5,
  mu = rep(0, cols),
  rows_test = 1000,
  seed = 1,
  varsN = NULL,
  varsB = NULL,
  varsE = NULL,
  varsP = NULL
)
}
\arguments{
\item{rows_train}{number of rows of the training sample. Default is 50,
for an high-dimensional setting}

\item{cols}{number of variables of both the training and the test samples.
Default is 2000, for an high-dimensional setting}

\item{cov}{covariance in the covariace matrix of the normal distribution.
Increasing it, the created features are more correlated. Default is 0.5}

\item{mu}{mean of the (multivariate) normal distribution. Default is 0}

\item{rows_test}{number of rows of the test sample. Default is 1000.
We suggest to create a test sample much bigger than the training sample
to test your method/model on more data}

\item{seed}{fix the seed. Default is 1}

\item{varsN}{select 2 variables of the first cols/4 variables to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varsB}{select 2 variables of the second cols/4 variables to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varsE}{select 2 variables of the third cols/4 variables to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varsP}{select 2 variables of the forth cols/4 variables to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}
}
\value{
a list of elements containing the created dataframe in
different versions (scaled and not) and additional information
}
\description{
function to create a synthetic sample with a target variable.
all the created variables follow a normal distribution
}
\examples{
library(pye)
df <- create_sample(cols = 20, rows_test = 10,
varsN = c(2, 4), varsB = c(6, 8), varsE = c(12, 14), varsP = c(16, 18))

}
