% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_dataset_creation.R
\name{create_sample_with_covariates}
\alias{create_sample_with_covariates}
\title{create_sample_with_covariates}
\usage{
create_sample_with_covariates(
  rows_train = 50,
  cols = 2000,
  cols_cov = 20,
  covar = 0.3,
  covar_cov = 0.5,
  mu = rep(0, cols),
  mu_cov = rep(0, cols_cov),
  rows_test = 1000,
  seed = 1,
  varsN = NULL,
  varsB = NULL,
  varsE = NULL,
  varsP = NULL,
  varN_cov = NULL,
  varB_cov = NULL,
  varE_cov = NULL,
  varP_cov = NULL
)
}
\arguments{
\item{rows_train}{number of rows of the training sample. Default is 50,
for an high-dimensional setting}

\item{cols}{number of regressor variables of both the training and the
test samples. Default is 2000, for an high-dimensional setting}

\item{cols_cov}{number of covariate variables of both the training and the
test samples. Default is 20, increase it for an high-dimensional setting}

\item{covar}{covariance in the covariace matrix of the normal distribution.
Increasing it, the created features are more correlated. Default is 0.5}

\item{mu}{mean of the (multivariate) normal distribution of the regressors.
Default is 0}

\item{mu_cov}{mean of the (multivariate) normal distribution of the
covariates. Default is 0}

\item{rows_test}{number of rows of the test sample. Default is 1000.
We suggest to create a test sample much bigger than the training sample
to test your method/model on more data}

\item{seed}{fix the seed. Default is 1}

\item{varsN}{select 2 variables of the first cols/4 regressors to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varsB}{select 2 variables of the second cols/4 regressors to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varsE}{select 2 variables of the third cols/4 regressors to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varsP}{select 2 variables of the forth cols/4 regressors to be
real regressors, combined with the others for the creation of the
synthetic target variable. Default is NULL}

\item{varN_cov}{select 1 variable of the first cols_cov/4 covariates
to be real covariates, combined with the others for the creation of
the synthetic target cut-off. Default is NULL}

\item{varB_cov}{select 1 variable of the second cols_cov/4 covariates
to be real covariates, combined with the others for the creation of
the synthetic target cut-off. Default is NULL}

\item{varE_cov}{select 1 variable of the third cols_cov/4 covariates
to be real covariates, combined with the others for the creation of
the synthetic target cut-off. Default is NULL}

\item{varP_cov}{select 1 variable of the forth cols_cov/4 covariates
to be real covariates, combined with the others for the creation of
the synthetic target cut-off. Default is NULL}
}
\value{
a list of elements containing the created dataframe in
different versions (scaled and not) and additional information
}
\description{
function to create a synthetic sample with a target variable.
This function creates 2 kind of variables: regressors and covariates,
setting useful to test, among others, pye, covYI and cPYE. The created
regressors follow a normal distribution while some of created covariates
follow a normal distribution and some others a bernoulli distribution.
}
\examples{
library(pye)
df <- create_sample_with_covariates(cols = 20, rows_test = 10,
  varsN = c(2, 4), varsB = c(6, 8), varsE = c(12, 14), varsP = c(16, 18))

}
