% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCP_penalty.R
\name{MCP_function}
\alias{MCP_function}
\title{MCP_function}
\usage{
MCP_function(betas, lambda, a)
}
\arguments{
\item{betas}{vertor to apply the proximal operator}

\item{lambda}{the penalization parameter}

\item{a}{the hyperparameter of the penalization MCP}
}
\value{
the value of the penalty MCP function
}
\description{
MCP penalization function
}
\examples{
library(pye)

betas <- seq(0, 2, by=0.2)
lambda <- 0.5
a <- 3.0
MCP_function(betas=betas, lambda=lambda, a=a)

}
