% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_proximal_operators.R
\name{proximal_operator_EN}
\alias{proximal_operator_EN}
\title{proximal_operator_EN}
\usage{
proximal_operator_EN(betas, lambda, alpha, ...)
}
\arguments{
\item{betas}{vertor to apply the proximal operator}

\item{lambda}{the penalization parameter}

\item{alpha}{the perameter of the Elastic-Net penalty to weight
the penalization L1 with respect of L2}

\item{...}{possibility to insert additional variables if 
necessary, not affecting the result}
}
\description{
Proximal operator related to the penalty Elastic-Net.
The "backward" step of the "forward-backward splitting method",
also named Proximal methods, used to optimize pye.
}
\examples{
library(pye)

betas <- seq(0, 2, by=0.2)
lambda <- 0.5
alpha <- 0.5
proximal_operator_EN(betas=betas, lambda=lambda, alpha=alpha)

}
