% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pye_proximal_operators.R
\name{proximal_operator_L12}
\alias{proximal_operator_L12}
\title{proximal_operator_L12}
\usage{
proximal_operator_L12(betas, lambda, ...)
}
\arguments{
\item{betas}{vertor to apply the proximal operator}

\item{lambda}{the penalization parameter}

\item{...}{possibility to insert additional variables if 
necessary, not affecting the result}
}
\description{
Proximal operator related to the penalty L1/2.
The "backward" step of the "forward-backward splitting method",
also named Proximal methods, used to optimize pye.
}
\examples{
library(pye)

betas <- seq(0, 2, by=0.2)
lambda <- 0.5
proximal_operator_L12(betas=betas, lambda=lambda)

}
