% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAD_penalty.R
\name{SCAD_function}
\alias{SCAD_function}
\title{SCAD_function}
\usage{
SCAD_function(betas, lambda, a)
}
\arguments{
\item{betas}{vertor to apply the proximal operator}

\item{lambda}{the penalization parameter}

\item{a}{the hyperparameter of the penalization SCAD}
}
\value{
the value of the penalty SCAD function
}
\description{
SCAD penalization function
}
\examples{
library(pye)

betas <- seq(0, 2, by=0.2)
lambda <- 0.5
a <- 3.7
SCAD_function(betas=betas, lambda=lambda, a=a)

}
