% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_df_for_pye.R
\name{scaling_df_for_pye}
\alias{scaling_df_for_pye}
\title{scaling_df_for_pye}
\usage{
scaling_df_for_pye(df, X, y)
}
\arguments{
\item{df}{the input dataset}

\item{y}{the target variable, to not scale (we are in a classification
environment)}

\item{n_folds}{number of fold of the cross validation}

\item{x}{regressors to be scaled. The function is not applied to the
intercept, dummys and non-regressors variables.}
}
\value{
the scaled database
}
\description{
function to scale and center variables, not applied to the
intercept, dummys and non-regressors variables.
}
\examples{
library(pye)
cols <- 2000
cols_cov <- 20
seed=1
simMicroarrayData_cov02_dim50_covariates <- create_sample_with_covariates(
		rows_train=50, cols=cols, cols_cov=cols_cov, covar=0.2, seed=seed)
df <- simMicroarrayData_cov02_dim50_covariates$train_df_scaled
X <- simMicroarrayData_cov02_dim50_covariates$X
y <- simMicroarrayData_cov02_dim50_covariates$y
C <- simMicroarrayData_cov02_dim50_covariates$C

df <- scaling_df_for_pye (df=df, X=colnames(df[, names(df) \%in\% c(X,C)]), y="y")
print(df)

}
